moved tests/util -> tests/common
